#15949: fix signed/unsigned comparison warning in importpdf.cpp
git-svn-id: svn://scribus.net/trunk/Scribus@23364 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/import/pdf/importpdf.cpp
@@ -633,7 +633,7 @@ bool PdfPlug::convert(const QString& fn)
 						}
 						m_Doc->setPageSize("Custom");
 					//	m_Doc->pdfOptions().PresentVals.clear();
-						for (int i = 0; i < pageNs.size(); ++i)
+						for (size_t i = 0; i < pageNs.size(); ++i)
 						{
 							int pp = pageNs[i];
 							m_Doc->setActiveLayer(baseLayer);
